// src/app/interfaces/otplanta.ts
export interface OtPlanta{
    id?: number;
    folio: number;	
    obra: string;
    area: string;	
    camion: string;	
    fecha: string;	
    nombre_operador: string;	
    num_trabajador: string;	
    consumo_diesel: string;	
    consumo_aceite_motor: string;	
    consumo_aceite_transmision: string;	
    consumo_aceite_hidraulico: string;	
    turno: string;	
    ubicacion_inicial: string;	
    h_horainicial?: string;
    h_horafinal?: string;
    kilometraje_i: string;	
    kilometraje_f: string;	
    ubicacion_final: string;	
    codigo_operacion: string;	
    nivel: string;	
    lugar: string;	
    empresa: string;	
    horainicial: string;	
    horafinal: string;	
    horometro_inicial1?: string;	
    horometro_final1?: string;	
    observaciones: string;	
    codigo_operacion2: string;	
    nivel2: string;	
    lugar2: string;	
    empresa2: string;	
    horainicial2: string;	
    horafinal2: string;	
    horometro_inicial2: string;	
    horometro_final2: string;	
    observaciones2: string;	
    codigo_operacion3: string;	
    nivel3: string;	
    lugar3: string;	
    empresa3: string;	
    horainicial3: string;	
    horafinal3: string;	
    horometro_inicial3: string;	
    horometro_final3: string;	
    observaciones3: string;	
    codigo_operacion4: string;	
    nivel4: string;	
    lugar4: string;	
    empresa4: string;	
    horainicial4: string;	
    horafinal4: string;
    horometro_inicial4: string;	
    horometro_final4: string;	
    observaciones4: string;	
    codigo_operacion5: string;	
    nivel5: string;	
    lugar5: string;	
    empresa5: string;	
    horainicial5: string;	
    horafinal5: string;	
    horometro_inicial5: string;	
    horometro_final5: string;	
    observaciones5: string;	
    codigo_operacion6: string;
    nivel6: string;	
    lugar6: string;	
    empresa6: string;	
    horainicial6: string;	
    horafinal6: string;	
    horometro_inicial6: string;	
    horometro_final6: string;	
    observaciones6: string;	
    codigo_operacion7: string;	
    nivel7: string;	
    lugar7: string;	
    empresa7: string;	
    horainicial7: string;	
    horafinal7: string;	
    horometro_inicial7: string;	
    horometro_final7: string;	
    observaciones7: string;	
    codigo_operacion8: string;	
    nivel8: string;	
    lugar8: string;	
    empresa8: string;	
    horainicial8: string;	
    horafinal8: string;	
    horometro_inicial8: string;	
    horometro_final8: string;	
    observaciones8: string;	
    ob_generales: string;
    nombre_supavar: string;	
    nombre_supcontratista: string;	
    created_at?: string;	
    updated_at?: string;	
}